rename stitcher!
