+++ tile_stitcher/__init__.py
@@ -1,3 +1,3 @@
-from .sticher import DATASET_SHORTNAMES, get_raster_from_tiles
+from .stitcher import DATASET_SHORTNAMES, get_raster_from_tiles
 
 __all__ = ['get_raster_from_tiles', 'DATASET_SHORTNAMES']
