+++ tile_stitcher/stitcher.py
